app/requirements.txt
#required libraries
aiohttp==3.6.2
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
uvicorn==0.11.3
pydicom==2.3.0
